Turn hints server into module
